third_party/blink/web_tests/http/tests/security/powerfulFeatureRestrictions/device-orientation-events-unavailable-on-insecure-origins.html

<!DOCTYPE html>
<title>Device Sensor Events not exposed to insecure origins</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script>

if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
  window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
} else {

  test(() => {
    assert_false('DeviceMotionEvent' in window);
    assert_false('DeviceOrientationEvent' in window);
    assert_false('DeviceOrientationAbsoluteEvent' in window);
    assert_false('DeviceMotionEventAcceleration' in window);
    assert_false('DeviceMotionEventRotationRate' in window);
    assert_false('ondevicemotion' in window);
    assert_false('ondeviceorientation' in window);
    assert_false('ondeviceorientationabsolute' in window);
  }, 'Event interfaces and event handlers are not exposed on `window`.');
}

</script>
